Basic chat application
This project encompasses a basic chat application developed in the Kotlin programming language. The application, built using the Firebase framework, allows users to register with their email and password, and subsequently log in to send text messages to the other party. Its fundamental functions include users registering with their email and password, followed by logging in to send text messages. The sent messages and user information are securely stored in the Firebase database. This project illustrates the development process of a basic chat application using the Kotlin language and ensures data security through Firebase integration.